Add caching mechanism for workspace packages and improve file modification time handling

+++ functions/_monorepo_get_workspace_packages.fish
@@ -1,4 +1,29 @@
 function _monorepo_get_workspace_packages
+    # Create tmp cache directory based on current working directory
+    set -l safe_dir (string replace -a '/' '_' $PWD)
+    set -l cache_dir /tmp/monorepo_cache/$safe_dir
+
+    if test -f "./package.json"
+        set -f pwd_hash (_monorepo_mtime_hash (git ls-files '*package.json'))
+    end
+
+    if test -f "./Cargo.toml"
+        set -f cargo_hash (_monorepo_mtime_hash (git ls-files '*Cargo.toml'))
+        if test -n "$pwd_hash"
+            set -f pwd_hash (string join - $pwd_hash $cargo_hash | sha256sum | awk '{print $1}')
+        else
+            set -f pwd_hash $cargo_hash
+        end
+    end
+
+    set -l cache_file "$cache_dir/$pwd_hash.json"
+
+    if test -f $cache_file
+        # Just use the cache file if it exists
+        cat $cache_file
+        return 0
+    end
+
     set -l packages_file (mktemp)
 
     echo "[]" >$packages_file
@@ -21,6 +46,7 @@ function _monorepo_get_workspace_packages
         return 1
     end
 
-    cat $packages_file
+    mkdir -p "$cache_dir"
+    cat $packages_file | tee "$cache_file"
     rm $packages_file
 end
